# swagger.yml
openapi: 3.0.0
info:
  title: Flask 应用程序 API
  description: 这是一个用于验证 iOS 应用内购买并处理 Apple 通知的 API 文档。
  version: 1.0.0
servers:
  - url: http://{host}:{port}
    description: Development server
    variables:
      host:
        default: localhost
      port:
        default: 5000
paths:
  /verify_purchase:
    post:
      summary: 验证 iOS 应用内购买收据并记录用户订阅信息
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                receipt:
                  type: string
                  description: iOS 应用内购买的收据数据
                  example: "receipt_data"
                user_id:
                  type: string
                  description: 用户的唯一标识符
                  example: "user_123"
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      product_id:
                        type: string
                        example: "com.example.product1"
                      purchase_date:
                        type: string
                        format: date-time
                        example: "2023-01-01T00:00:00"
                      expiry_date:
                        type: string
                        format: date-time
                        example: "2024-01-01T00:00:00"
        '400':
          description: 请求错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "No receipt or user_id provided"
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"
  /apple_notifications:
    post:
      summary: 处理 Apple 的通知
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                notification_type:
                  type: string
                  description: 通知类型
                  example: "INITIAL_BUY"
                user_id:
                  type: string
                  description: 用户的唯一标识符
                  example: "user_123"
                product_id:
                  type: string
                  description: 购买的产品 ID
                  example: "com.example.product1"
                purchase_date:
                  type: string
                  format: date-time
                  description: 购买日期和时间
                  example: "2023-01-01T00:00:00"
                expiry_date:
                  type: string
                  format: date-time
                  description: 订阅到期日期和时间
                  example: "2024-01-01T00:00:00"
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '400':
          description: 请求错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid notification data"
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"
components:
  schemas:
    Subscription:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: string
          example: "user_123"
        product_id:
          type: string
          example: "com.example.product1"
        purchase_date:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00"
        expiry_date:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00"
        status:
          type: string
          example: "active"